tests: Move test to mutex_test.go
